+++ src/init.c
@@ -1,0 +1,18 @@
+#include <R.h>
+#include <Rinternals.h>
+#include <stdlib.h> // for NULL
+#include <R_ext/Rdynload.h>
+
+/* .Call calls */
+extern SEXP base64_enc(SEXP);
+
+static const R_CallMethodDef CallEntries[] = {
+  {"base64_enc", (DL_FUNC) &base64_enc, 1},
+  {NULL, NULL, 0}
+};
+
+void R_init_xfun(DllInfo *dll)
+{
+  R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
+  R_useDynamicSymbols(dll, FALSE);
+}